requirements.txt
streamlit
pyyaml
bcrypt
PyJWT
extra-streamlit-components
pymongo
deta